small design fixes
