+++ app/_layout.tsx
@@ -1,7 +1,7 @@
 import { Slot } from 'expo-router'
 import { GluestackUIProvider } from '@gluestack-ui/themed'
 import { config } from '@gluestack-ui/config'
-import DrawerComponent from '../components/DrawerComponent';
+import DrawerComponent from '../components/Drawer';
 import ProductProvider from '../hooks/productDetails';
 import { StatusBar } from 'react-native';
 import CategoryProvider from '../hooks/categoryDetails';
