+++ app/portfolio.tsx
@@ -1,17 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { Text, View } from 'react-native';
-import ImageComponent from '../components/ImageComponent';
-import ProductItemComponent from '../components/ProductItemComponent';
+import { View } from 'react-native';
 import { ScrollView } from '@gluestack-ui/themed';
-import TitleComponent from '../components/TitleComponent';
-import BestProductComponent from '../components/BestProductComponent';
 import { router } from 'expo-router';
 import { useProduct } from '../hooks/productDetails';
+import TitleComponent from '../components/TitleComponent';
+import BestProduct from '../components/BestProduct';
+import ProductItem from '../components/ProductItem';
 
 export default function Home() {
-
     const product = useProduct()
-    
     interface IProductItem {
         id: number,
         name: string,
@@ -23,7 +20,6 @@ export default function Home() {
     }
 
     const [list, setList] = useState<IProductItem[]>([])
-
     useEffect(() =>{
         fetch('https://api-catalogo-pi.onrender.com/product')
         .then(response => response.json())
@@ -31,41 +27,40 @@ export default function Home() {
     })
 
     return (
-            <ScrollView>
-                <TitleComponent title={'Portifólio Completo'}></TitleComponent>
-                <View style={{ alignItems: "center" }}>
-                {list.map((item, index) => {
-                    if (index === 0) {
-                        return (
-                            <BestProductComponent
-                                key={index}
-                                id={item.id}
-                                name={item.name}
-                                price={'R$ ' + item.price.toFixed(2)}
-                                photo={item.photo1}
-                                promotion={item.promotion}
-                                description={item.description} 
-                                onPress={() => {product.setProductId({id: item.id})
-                                router.push('/productDetail')
-                            }}/>
-                        );
-                    } else {
-                        return (
-                            <ProductItemComponent
-                                key={index}
-                                id={item.id}
-                                name={item.name}
-                                price={'R$ ' + item.price.toFixed(2)}
-                                photo={item.photo1}
-                                promotion={item.promotion}
-                                onPress={() => {product.setProductId({id: item.id})
-                                router.push('/productDetail')
-                            }}/>
-                        );
-                    }
-                })}
+        <ScrollView>
+            <TitleComponent title={'Portifólio Completo'}></TitleComponent>
+            <View style={{ alignItems: "center" }}>
+            {list.map((item, index) => {
+                if (index === 0) {
+                    return (
+                        <BestProduct
+                            key={index}
+                            id={item.id}
+                            name={item.name}
+                            price={'R$ ' + item.price.toFixed(2)}
+                            photo={item.photo1}
+                            promotion={item.promotion}
+                            description={item.description} 
+                            onPress={() => {product.setProductId({id: item.id})
+                            router.push('/productDetail')
+                        }}/>
+                    );
+                } else {
+                    return (
+                        <ProductItem
+                            key={index}
+                            id={item.id}
+                            name={item.name}
+                            price={'R$ ' + item.price.toFixed(2)}
+                            photo={item.photo1}
+                            promotion={item.promotion}
+                            onPress={() => {product.setProductId({id: item.id})
+                            router.push('/productDetail')
+                        }}/>
+                    );
+                }
+            })}
             </View>
-            </ScrollView>
+        </ScrollView>
     );
-    
 }
