+++ app/scentDetail0.tsx
