+++ app/scentDetail1.tsx
