+++ app/scentDetail3.tsx
