+++ components/ImageComponent.tsx
@@ -1,38 +1,36 @@
-import { Link } from "expo-router";
 import { Text, View, Image, StyleSheet, ImageSourcePropType } from "react-native";
 
-interface IImage {
+interface ImageComponentProps {
     imageUrl: ImageSourcePropType | undefined
     descrição: string;
     descrição1: string;
 }
 
-export default function ImageComponent({imageUrl, descrição, descrição1}:IImage) {
+export default function ImageComponent({imageUrl, descrição, descrição1}:ImageComponentProps) {
     return(
         <View  style={styles.container}>
             <Image
                 style={styles.stretch}
                 source={imageUrl}
             />
-             <Text style={styles.text}>{descrição}</Text>
-             {/* {'\n'} */}
-             <Text style={styles.text}>{descrição1}</Text>
+            <Text style={styles.text}>{descrição}</Text>
+            <Text style={styles.text}>{descrição1}</Text>
         </View>
     )
 }
 const styles = StyleSheet.create({
     stretch: {
         marginLeft: '5%',
-      width: 200,
-      height: 200,
-      resizeMode: 'stretch',
+        width: 200,
+        height: 200,
+        resizeMode: 'stretch',
     },
     container: {
-        flexDirection: 'row', // Estabelece o layout em linha (horizontal)
-        alignItems: 'center', // Alinha os itens verticalmente ao centro
+        flexDirection: 'row',
+        alignItems: 'center',
         paddingTop: 50,
-      },
-      text: {
-        marginLeft: 10, // Adiciona um espaço à esquerda do texto para separá-lo da imagem
-      },
-  });
+    },
+    text: {
+        marginLeft: 10,
+    },
+});
