+++ components/LoginButtonPressable.tsx
@@ -8,15 +8,12 @@ interface LoginButtonPressableProps {
 
 export default function LoginButtonPressable({title, onPress, color}: LoginButtonPressableProps){
     return(
-        // crindo nosso componente
-        <>
-            <Pressable 
-                style={styles.button}
-                onPress={onPress}
-            >
-                <Text style={styles.text}>{title}</Text>
-            </Pressable>
-        </>
+        <Pressable 
+            style={styles.button}
+            onPress={onPress}
+        >
+            <Text style={styles.text}>{title}</Text>
+        </Pressable>
     )
 }
 
@@ -38,6 +35,5 @@ const styles = StyleSheet.create({
         fontWeight: 'bold',
         letterSpacing: 0.25,
         color: 'white',
-      },
-    
+    },
 })
